app/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
PyJWT==2.3.0

